# GameShop/Dto/ProducerDto.cs
using GameShop.Models;

namespace GameShop.Dto
{
    public class ProducerDto
    {
        public int Id { get; set; }
        public string ProducerName { get; set; }
        public string ProducerDescrip { get; set; }
        public string ProducerCountry { get; set; }
        //public ICollection<Game> Games { get; set; }
    }
}