# GameShop/Models/Producer.cs
namespace GameShop.Models
{
    public class Producer
    {
        public int Id { get; set; }
        public string ProducerName { get; set; }
        public string ProducerDescrip { get; set; }
        public string ProducerCountry { get; set; }
        //public ICollection<Game> Games { get; set; }
    }
}